Remove ambiguity in ichat-gw.cxx, minor bugs in IPCThread.cxx and Server.cxx

+++ apps/ichat-gw/Server.cxx
@@ -751,7 +751,7 @@ Server::checkSubscription(const std::string& to, const std::string& from)
    msg.addArg("sendSubscriptionResponse");
    msg.addArg(from.c_str());
    msg.addArg(to.c_str());   
-   msg.addArg(Data((unsigned long)success).c_str());   
+   msg.addArg(Data((UInt32)success).c_str());   
    mIPCThread->sendIPCMsg(msg);
 }
